requirements-gmp.txt
flask
flask-cors
flask-sqlalchemy
sqlalchemy
python-dotenv
requests
beautifulsoup4
python-docx
openpyxl
pandas
reportlab
PyPDF2
pdfplumber
pydantic
lxml
gunicorn
